backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django-cors-headers>=4.9.0",
    "django-debug-toolbar>=6.0.0",
    "djangorestframework>=3.16.1",
    "pyautogui>=0.9.54",
    "pyvjoy>=1.0.1",
]
